feat: break down util.py
