+++ src/rabbitmq_export/rabbitmq_export/export.py
@@ -4,7 +4,7 @@ import requests
 
 import rabbitmq_export.context
 
-from lib.util import (
+from lib.export_backup import (
   backup_to_s3,
   setup_local_export_dir,
   write_file
